Do not overwrite /etc/rc.local.bak in case shrinking a previously shrunken image that has not yet run the expansion rc.local
If you pishrink an image, it saves the original /etc/rc.local as /etc/rc.local.bak. Upon re-expansion, the backup is restored.

If you take that shrunken image and expand it, make changes, and pishrink again, /etc/rc.local.bak gets overwritten with the "re-expand and restore backup" /etc/rc.local from the first pishrink. When that image eventually boots, it causes an rc.local loop.

This commit changes it to leave an existing /etc/rc.local.bak as it is.

+++ pishrink.sh
@@ -86,7 +86,7 @@ function set_autoexpand() {
 
     if [[ -f "$mountdir/etc/rc.local" ]] && [[ "$(md5sum "$mountdir/etc/rc.local" | cut -d ' ' -f 1)" != "1c579c7d5b4292fd948399b6ece39009" ]]; then
       echo "Creating new /etc/rc.local"
-    if [ -f "$mountdir/etc/rc.local" ]; then
+    if [ -f "$mountdir/etc/rc.local" -a ! -f "$mountdir/etc/rc.local.bak"]; then
         mv "$mountdir/etc/rc.local" "$mountdir/etc/rc.local.bak"
     fi
